Split strict exact proof gates
